Bus-based implementation
Uses a bus to simplify the connection of information being transported between blocks.  Currently implements the sensory state, neural control and biomechanics as separate matlab function blocks.  Next step is to replace biomechanics block with the pure simulink implementation.

The bus definition and initialization are defined in the class defined in "AplysiaSimulink.m"

The simulink implementation is in "Test.slx"

A file that instantiates the class, runs the simulink simulation and generates the plot is found in "TestAplysiaClass.m"